LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-animation-with-scroll-timeline-in-hidden-subtree-expected.txt


FAIL Animation with scroll-timeline should be affected c-v assert_approx_equals: scrollContainer scrolls to bottom, so the opacity should be 0 expected 0 +/- 0.1 but got 1
